--- MoonDawgArduino/Libraries/Movement/Movement.properties
@@ -1,17 +1,0 @@
-name=MovementMoonDawg
-
-version=1.5
-
-author=Cameron Schwartzberg
-
-maintainer=camhotwheels@gmail.com
-
-sentence=This contains all of the movement classes and functions for the MoonDawgArduinoCode.ino.
-
-paragraph= This code includes a header file named `Movement.h`, which serves as the interface for the corresponding `Movement.cpp` file. Authored by Cameron Schwartzberg, Marco Caliendo, Andrew Barnes, Nadia Lafontant, and Mitchell Wettig, the header file contains class declarations for `DigMovement`, `DepositMovement`, and `DriveMovement`. Each class is equipped with private `Servo` motor instances, `motor1` and `motor2`, and public methods to control their movements. The `DigMovement` class allows control over a digging motor and an actuator, while the `DepositMovement` class manages a deposit motor and a vibrator. The `DriveMovement` class is designed for drive motors. All classes have an empty constructor for initialization and methods to set up the motor pins, set initial positions, and control their movements using the `write()` method from the `Arduino` library. Motor speeds are represented by values of 0 (backward), 90 (stop), and 180 (forward). The header file also includes necessary libraries such as `stdint.h` and `Servo.h`, along with the `Arduino.h` library. This structured approach aims to provide a modular and organized system for controlling various motors and actuators in robotic or automated applications.
-
-category=Classes
-
-url=https://github.com/SIU-Robotics/embedded-berm-bot
-
-architectures=*
