+++ SandstormArduino/Libraries/CameraControl/Camera.cpp
