+++ SandstormArduino/Libraries/CameraControl/Camera.h
