+++ SandstormArduino/Libraries/PinDefinitions/PinDefinitions.properties
@@ -1,0 +1,15 @@
+name=PinDefinitionsMoonDawg
+
+version=1.3.0
+
+author=Cameron Schwartzberg
+
+maintainer=camhotwheels@gmail.com
+
+sentence=This contains all of the pin definitions for the MoonDawgArduinoCode.ino.
+
+category=Definitions
+
+url=https://github.com/SIU-Robotics/embedded-berm-bot
+
+architectures=*
